Address Matteo's comments and fix spotbugs
