+++ web/src/main/bvmui/src/views/SystemStatus.vue
