+++ web/src/main/java/org/bookkeepervisualmanager/api/resources/LedgersResource.java
@@ -19,13 +19,9 @@
  */
 package org.bookkeepervisualmanager.api.resources;
 
-import java.io.Serializable;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Map.Entry;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -93,161 +89,6 @@ public class LedgersResource extends AbstractBookkeeperResource {
         return b;
     }
 
-    public static final class LedgerBean implements Serializable {
-
-        private long id;
-        private Map<String, String> metadata = new HashMap<>();
-        private long age;
-        private int ensembleSize;
-        private int writeQuorumSize;
-        private int ackQuorumSize;
-        private long lastEntryId;
-        private long length;
-        private String password;
-        private String digestType;
-        private long ctime;
-        private boolean closed;
-        private String state;
-        private int metadataFormatVersion;
-        private List<String> bookies;
-
-        public long getAge() {
-            return age;
-        }
-
-        public void setAge(long age) {
-            this.age = age;
-        }
-
-        public long getId() {
-            return id;
-        }
-
-        public void setId(long id) {
-            this.id = id;
-        }
-
-        public Map<String, String> getMetadata() {
-            return metadata;
-        }
-
-        public void setMetadata(Map<String, String> metadata) {
-            this.metadata = metadata;
-        }
-
-        public void setMetadataValue(String key, byte[] value)  {
-            this.metadata.put(key, new String(value, StandardCharsets.UTF_8));
-        }
-
-        public void setLedgerMetadata(LedgerMetadata metadata)  {
-            Map<String, byte[]> customMetadata = metadata.getCustomMetadata();
-            for (Entry<String, byte[]> currentCustomMetadata : customMetadata.entrySet()) {
-                setMetadataValue(currentCustomMetadata.getKey(), currentCustomMetadata.getValue());
-            }
-        }
-
-        public int getEnsembleSize() {
-            return ensembleSize;
-        }
-
-        public void setEnsembleSize(int ensembleSize) {
-            this.ensembleSize = ensembleSize;
-        }
-
-        public int getWriteQuorumSize() {
-            return writeQuorumSize;
-        }
-
-        public void setWriteQuorumSize(int writeQuorumSize) {
-            this.writeQuorumSize = writeQuorumSize;
-        }
-
-        public int getAckQuorumSize() {
-            return ackQuorumSize;
-        }
-
-        public void setAckQuorumSize(int ackQuorumSize) {
-            this.ackQuorumSize = ackQuorumSize;
-        }
-
-        public long getLastEntryId() {
-            return lastEntryId;
-        }
-
-        public void setLastEntryId(long lastEntryId) {
-            this.lastEntryId = lastEntryId;
-        }
-
-        public long getLength() {
-            return length;
-        }
-
-        public void setLength(long length) {
-            this.length = length;
-        }
-
-        public String getPassword() {
-            return password;
-        }
-
-        public void setPassword(String password) {
-            this.password = password;
-        }
-
-        public String getDigestType() {
-            return digestType;
-        }
-
-        public void setDigestType(String digestType) {
-            this.digestType = digestType;
-        }
-
-        public long getCtime() {
-            return ctime;
-        }
-
-        public void setCtime(long ctime) {
-            this.ctime = ctime;
-        }
-
-        public boolean isClosed() {
-            return closed;
-        }
-
-        public void setClosed(boolean closed) {
-            this.closed = closed;
-        }
-
-        public String getState() {
-            return state;
-        }
-
-        public void setState(String state) {
-            this.state = state;
-        }
-
-        public int getMetadataFormatVersion() {
-            return metadataFormatVersion;
-        }
-
-        public void setMetadataFormatVersion(int metadataFormatVersion) {
-            this.metadataFormatVersion = metadataFormatVersion;
-        }
-
-        public List<String> getBookies() {
-            return bookies;
-        }
-
-        public void setBookies(List<String> bookies) {
-            this.bookies = bookies;
-        }
-
-        @Override
-        public String toString() {
-            return "LedgerBean{" + "id=" + id + ", metadata=" + metadata + '}';
-        }
-
-    }
 
     private Integer convertParam(String s) {
         if (s == null || s.trim().isEmpty()) {
