+++ web/src/main/java/org/bookkeepervisualmanager/api/resources/SystemStatusResource.java
@@ -114,73 +114,6 @@ public class SystemStatusResource extends AbstractBookkeeperResource {
 
     @GET
     @Secured
-    @Path("clusterconfiguration")
-    @Produces(MediaType.APPLICATION_JSON)
-    public ClusterWideConfiguration getClusterWideConfiguration() throws Exception {
-        BookkeeperManager.RefreshCacheWorkerStatus refreshWorkerStatus = getBookkeeperManger().getRefreshWorkerStatus();
-        BookkeeperManager.ClusterWideConfiguration clusterWideConfiguration = refreshWorkerStatus.getLastClusterWideConfiguration();
-        return new ClusterWideConfiguration(clusterWideConfiguration.getAuditor(),
-                clusterWideConfiguration.isAutorecoveryEnabled(),
-                clusterWideConfiguration.getLostBookieRecoveryDelay(),
-                clusterWideConfiguration.getLayoutFormatVersion(),
-                clusterWideConfiguration.getLayoutManagerFactoryClass(),
-                clusterWideConfiguration.getLayoutManagerVersion(),
-                refreshWorkerStatus.getBookkkeeperClientConfiguration());
-    }
-
-    public static final class ClusterWideConfiguration {
-
-        private final String auditor;
-        private final boolean autorecoveryEnabled;
-        private final int lostBookieRecoveryDelay;
-        private final int layoutFormatVersion;
-        private final String layoutManagerFactoryClass;
-        private final int layoutManagerVersion;
-        private final String clientConfiguration;
-
-        public ClusterWideConfiguration(String auditor, boolean autorecoveryEnabled, int lostBookieRecoveryDelay, int layoutFormatVersion, String layoutManagerFactoryClass, int layoutManagerVersion,
-                String clientConfiguration) {
-            this.auditor = auditor;
-            this.autorecoveryEnabled = autorecoveryEnabled;
-            this.lostBookieRecoveryDelay = lostBookieRecoveryDelay;
-            this.layoutFormatVersion = layoutFormatVersion;
-            this.layoutManagerFactoryClass = layoutManagerFactoryClass;
-            this.layoutManagerVersion = layoutManagerVersion;
-            this.clientConfiguration = clientConfiguration;
-        }
-
-        public String getAuditor() {
-            return auditor;
-        }
-
-        public boolean isAutorecoveryEnabled() {
-            return autorecoveryEnabled;
-        }
-
-        public int getLostBookieRecoveryDelay() {
-            return lostBookieRecoveryDelay;
-        }
-
-        public int getLayoutFormatVersion() {
-            return layoutFormatVersion;
-        }
-
-        public String getLayoutManagerFactoryClass() {
-            return layoutManagerFactoryClass;
-        }
-
-        public int getLayoutManagerVersion() {
-            return layoutManagerVersion;
-        }
-
-        public String getClientConfiguration() {
-            return clientConfiguration;
-        }
-
-    }
-
-    @GET
-    @Secured
     @Path("refresh")
     @Produces(MediaType.APPLICATION_JSON)
     public SystemStatus refresh() throws Exception {
